Add files via upload
Immerse yourself in the classic game of Checkers with this interactive and engaging project. Built using Python and a user-friendly interface, the Checkers Game project offers a digital platform to play the timeless two-player board game. Enjoy strategic moves, captures, and crowned pieces as you challenge friends or test your skills against a computer opponent. The project showcases a sleek design and intuitive gameplay, providing a nostalgic experience for Checkers enthusiasts and a fun introduction for newcomers to this beloved strategy game.